chore: restructure search components
