+++ apps/marginfi-v2-trading/src/components/common/search/index.ts
@@ -1,1 +1,2 @@
-export * from "./search";
+export * from "./search-button";
+export * from "./search-input";
